requirements.txt
Flask==2.3.3
requests==2.31.0
clickhouse-driver==0.2.6
paramiko==3.4.0
Flask-HTTPAuth==4.8.0
PyYAML==6.0.1
Flask-Session==0.5.0
